requirements/pip.txt
gymnasium [atari]
gymnasium [accept-rom-license]
gym-super-mario-bros
numpy
pillow
pynput
nes-py
